Migrate tag page to appdir
